Add glm Nuget package.
Add gml Nuget package. Old solution used a glm package that was downloaded from GitHub. With the Nuget package, the package is loaded automatically on any system that clones the repository. With the old repsoitory, the user would have to download the GitHub repository before the solution could be built.